requirements.txt
discord.py
aiohttp
websockets
chardet
asyncio
nekos.py
